Move compress object to utils.
